requirements/development.txt
# Include base requirements
-r base.txt

# Development server
django-extensions>=3.2

# Debugging
django-debug-toolbar>=4.2
ipython>=8.0

# Testing
pytest>=7.4
pytest-django>=4.5
pytest-cov>=4.1
factory-boy>=3.3

# Code quality
flake8>=6.1
black>=23.0
isort>=5.12

# Type checking
mypy>=1.5
django-stubs>=4.2
